Reorganize code structure
Flattened out the code structure and refactored
components from Component.tsx to Component/index.tsx
