+++ packages/pipeline-editor/src/CustomFormControls/StringArrayControl.tsx
