+++ packages/pipeline-editor/src/NodeTooltip/index.tsx
@@ -1,33 +1,14 @@
-/*
- * Copyright 2018-2020 IBM Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import React from "react";
-
 import { hasValue, toPrettyString } from "./utils";
 
-export interface IProperty {
-  label: string;
-  value: any;
-}
-
-interface IProps {
+interface Props {
   error?: string;
-  properties: IProperty[];
+  properties: {
+    label: string;
+    value: any;
+  }[];
 }
 
-export const NodeTooltip: React.FC<IProps> = ({ error, properties }) => {
+function NodeTooltip({ error, properties }: Props) {
   return (
     <div className="elyra-PipelineNodeTooltip">
       {error && (
@@ -46,4 +27,6 @@ export const NodeTooltip: React.FC<IProps> = ({ error, properties }) => {
         ))}
     </div>
   );
-};
+}
+
+export default NodeTooltip;
